pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ymse"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "lettermint",
    "loguru",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["ymse"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8",
    "ruff>=0.9",
]
